Fix reroll to not add dice and to return a result for all dice rolled.
The original module only return the highest number rolled, which is fine for rolling
traits and aptitudes. When rolling damage, however, you need all the dice.

This fix does not add an extra die to the result for each explosion. Nor does it
only return one number.

Adding an extra die as well as adding the value to the original roll means each
explosion added twice the number of sides to the total rolled i.e. an exploding
d6 adds 12 to the total result instead of 6.